refactor

+++ servlets/src/main/java/ru/netology/service/PostService.java
@@ -7,36 +7,38 @@ import ru.netology.repository.PostRepository;
 import java.util.List;
 
 public class PostService {
-  private final PostRepository repository;
+    private final PostRepository repository;
 
-  public PostService(PostRepository repository) {
-    this.repository = repository;
-  }
+    public PostService(PostRepository repository) {
+        this.repository = repository;
+    }
 
-  public List<Post> all() {
-    return repository.all();
-  }
+    public List<Post> all() {
+        return repository.all();
+    }
 
-  public Post getById(long id) {
-    return repository.getById(id).orElseThrow(NotFoundException::new);
-  }
+    public Post getById(long id) {
+        return repository.getById(id).orElseThrow(NotFoundException::new);
+    }
 
-  public Post save(Post post) {
-    return repository.save(post);
-  }
+    public Post save(Post post) {
+        return repository.save(post);
+    }
 
-  public Post update(Post post, String content) {
-    return repository.update(post, content);
-  }
+    public void update(Long id, String content) {
+        Post post = getById(id);
+        if (post != null) {
+            post.setContent(content);
+            repository.update(post);
+        }
+    }
 
-  public void delete(Post post) {
-    repository.delete(post);
-  }
+    public void delete(Long id) {
+        Post post = getById(id);
+        if (post != null) {
+            repository.delete(post);
+        }
+    }
 
-
-
-//  public void removeById(long id) {
-//    repository.removeById(id);
-//  }
 }
